requirements.txt
streamlit
pytesseract
pillow
pandas
openpyxl
